[RLM-529] Add pre-flight check to ensure all passwords are set
This change adds a pre-flight check to ensure all passwords are set
prior to RPC-MaaS being deployed. A new playbook was created which uses
the example password file to iterate and find if any of the passwords
are unset from the playbook runtime. If a required password is missing
the task will produce a failure which instructs the deployer to set the
required password before continuing.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

Issue: [RLM-529](https://rpc-openstack.atlassian.net/browse/RLM-529)
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>